@@ -1,123 +1,75 @@
 {
-  "name": "update_version",
+  "name": "twom",
   "version": "0.0.1",
   "description": "A Quasar Framework app",
   "productName": "Quasar App",
-  "author": "",
+  "author": "jedika <jekjedika@gmail.com>",
   "private": true,
   "scripts": {
-    "lint": "eslint --ext .js,.ts,.vue --ignore-path .gitignore ./",
-    "test": "echo \"No test specified\" && exit 0",
-    "rebuild": "electron-rebuild -f -w sqlite3",
-    "codegen": "graphql-codegen --config codegen.yml"
+    "lint": "eslint --ext .js,.ts,.vue ./",
+    "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
-    "@erdii/caddy-serverless-express": "^0.0.2",
-    "@quasar/extras": "^1.0.0",
+    "@quasar/extras": "^1.9.18",
     "@types/dotenv": "^8.2.0",
-    "@types/electron-devtools-installer": "^2.2.0",
     "@types/graphql": "^14.5.0",
-    "@vue/apollo-composable": "^4.0.0-alpha.10",
-    "@vue/apollo-util": "^4.0.0-alpha.6",
-    "@vue/composition-api": "^1.0.0-beta.8",
-    "apollo-boost": "^0.4.9",
-    "apollo-cache-inmemory": "^1.6.6",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-link-batch-http": "^1.2.14",
-    "apollo-link-http": "^1.5.17",
-    "apollo-server-express": "^2.15.1",
-    "asar": "^3.0.3",
-    "babel-plugin-graphql-tag": "^2.5.0",
+    "@vue/apollo-composable": "^4.0.0-alpha.12",
+    "@vue/apollo-option": "^4.0.0-alpha.11",
+    "@vue/apollo-util": "^4.0.0-alpha.11",
+    "apollo-link-context": "^1.0.20",
+    "babel-plugin-graphql-tag": "^3.2.0",
     "child_process": "^1.0.2",
-    "copy-webpack-plugin": "^6.0.3",
+    "copy-webpack-plugin": "^7.0.0",
+    "core-js": "^3.8.3",
     "dataloader": "^2.0.0",
+    "date-format-parse": "^0.2.6",
+    "docx": "^5.5.0",
     "dotenv": "^8.2.0",
-    "electron-is-dev": "^1.2.0",
-    "electron-store": "^5.2.0",
     "express": "^4.17.1",
-    "fetch": "^1.1.0",
-    "graphql": "^15.3.0",
+    "file-saver": "^2.0.5",
+    "graphql": "^15.5.0",
     "graphql-tag": "^2.11.0",
-    "graphql-tools": "^6.0.13",
-    "gyp": "^0.5.0",
+    "graphql-tools": "^7.0.3",
+    "html-docx-js-typescript": "^0.1.5",
+    "htmltag": "^0.6.5",
     "idle-session-timeout": "^1.0.1",
-    "is-electron": "^2.2.0",
-    "jsonwebtoken": "^8.5.1",
-    "needle": "^2.5.0",
-    "node": "^14.5.0",
-    "node-addon-api": "^3.0.0",
-    "node-crypt": "^2.0.0",
-    "node-fetch": "^2.6.0",
-    "node-gyp": "^7.0.0",
-    "node-pre-gyp": "^0.15.0",
-    "node-static": "^0.7.11",
-    "normalizr": "^3.6.0",
-    "offline-token-generator": "^1.0.2",
-    "quasar": "^1.12.13",
-    "reflect-metadata": "^0.1.13",
-    "rimraf": "^3.0.2",
-    "rxjs": "^6.6.0",
+    "multilingual-number-to-words": "^5.0.3",
+    "normalizr": "^3.6.1",
+    "officegen": "^0.6.4",
+    "pinia": "^0.1.0",
+    "quasar": "^2.0.0-beta.8",
     "simple-ajax-uploader": "^2.6.7",
-    "ts-sql-plugin": "^0.8.1",
-    "two-way-decorator": "^0.2.2",
-    "type-graphql": "^1.0.0-rc.3",
-    "typeorm": "^0.2.25",
-    "urllib": "^2.36.1",
-    "voyager": "^0.4.13",
-    "vue-apollo": "^3.0.4",
-    "vue-class-component": "^7.2.2",
-    "vue-i18n": "^8.18.2",
-    "vue-property-decorator": "^8.3.0",
-    "vuex-class": "^0.3.2",
-    "vuex-map-decorators": "^1.2.7",
-    "vuex-map-fields": "^1.4.0",
-    "vuex-persistedstate": "^3.1.0",
-    "webpack-html-plugin": "^0.1.1"
+    "subscriptions-transport-ws": "^0.9.18",
+    "tauri": "^0.14.1",
+    "vue-demi": "^0.6.0",
+    "vue-html-print": "^1.3.1",
+    "vue-html-to-paper": "^1.3.1",
+    "vue-print-nb": "^1.5.0"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-decorators": "^7.10.5",
-    "@graphql-codegen/cli": "^1.17.7",
-    "@graphql-codegen/introspection": "^1.17.7",
-    "@graphql-codegen/typescript": "^1.17.7",
-    "@graphql-codegen/typescript-operations": "^1.17.7",
-    "@graphql-codegen/typescript-vue-apollo": "^1.17.7",
-    "@nestjs/cli": "^7.0.0",
-    "@nestjs/schematics": "^7.0.0",
-    "@nestjs/testing": "^7.0.0",
-    "@quasar/app": "^2.0.0",
-    "@quasar/quasar-app-extension-typescript": "^1.0.2",
-    "@types/apollo-upload-client": "^8.1.3",
-    "@types/express": "^4.17.7",
-    "@types/jest": "^26.0.7",
-    "@types/node": "12.12.6",
-    "@types/supertest": "^2.0.8",
-    "@typescript-eslint/eslint-plugin": "^2.22.0",
-    "@typescript-eslint/parser": "^2.22.0",
-    "apollo-upload-client": "^14.1.0",
+    "@apollo/client": "^3.3.9",
+    "@graphql-codegen/cli": "^1.20.1",
+    "@graphql-codegen/introspection": "^1.18.1",
+    "@graphql-codegen/typescript": "^1.21.0",
+    "@graphql-codegen/typescript-operations": "^1.17.14",
+    "@graphql-codegen/typescript-vue-apollo": "^2.3.1",
+    "@nestjs/cli": "^7.5.4",
+    "@nestjs/schematics": "^7.2.7",
+    "@nestjs/testing": "^7.6.12",
+    "@quasar/app": "^3.0.0-beta.8",
+    "@types/apollo-upload-client": "^14.1.0",
+    "@types/express": "^4.17.11",
+    "@types/node": "^10.17.15",
+    "@typescript-eslint/eslint-plugin": "^4.9.0",
+    "@typescript-eslint/parser": "^4.9.0",
     "babel-eslint": "^10.0.1",
-    "core-js": "^3.6.5",
-    "devtron": "^1.4.0",
-    "electron": "^9.0.0",
-    "electron-builder": "^22.7.0",
-    "electron-debug": "^3.0.1",
-    "electron-devtools-installer": "^3.0.0",
-    "electron-packager": "^14.1.1",
-    "electron-rebuild": "^1.11.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-loader": "^3.0.3",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-vue": "^6.2.1",
-    "jest": "26.0.1",
-    "prettier": "^1.19.1",
-    "supertest": "^4.0.2",
-    "ts-jest": "26.1.0",
-    "ts-loader": "^6.2.1",
-    "ts-node": "^8.9.1",
-    "tsconfig-paths": "^3.9.0",
-    "workbox-webpack-plugin": "5.0.0"
+    "eslint": "^7.14.0",
+    "eslint-config-prettier": "^6.15.0",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-vue": "^7.0.0",
+    "ts-loader": "^8.0.17",
+    "ts-node": "^9.1.1",
+    "vue-apollo": "^3.0.0-alpha.3"
   },
   "browserslist": [
     "last 10 Chrome versions",
