@@ -1,23 +1,15 @@
 <template>
-  <div id="q-app">
-    <router-view />
-  </div>
+  <router-view />
 </template>
-
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-import isE from 'is-electron';
-const isElectron = isE();
-let remote;
-isElectron ? ({ remote } = require('electron')) : {};
+import { defineComponent, inject } from 'vue';
 
-@Component({ name: 'App' })
-export default class App extends Vue {
-  mounted() {
-    // console.log('user ***', await Query.login('jedikah', '123'));
-    // console.log({ app: await Query.users() });
-    console.log({ 'is electron': isElectron });
-    if (isElectron) console.log(remote);
-  }
-}
+export default defineComponent({
+  name: 'App',
+  setup() {
+    // const $htmlToPaper = inject('$htmlToPaper');
+
+    return {};
+  },
+});
 </script>
