@@ -9,12 +9,12 @@
  *
  * Boot files are your "main.js"
  **/
-import Vue from 'vue'
-import './import-quasar.js'
 
 
 
-import App from 'app/src/App.vue'
+import { Quasar } from 'quasar'
+import quasarUserOptions from './quasar-user-options.js'
+import RootComponent from 'app/src/App.vue'
 
 
 import createStore from 'app/src/store/index'
@@ -25,37 +25,35 @@ import createRouter from 'app/src/router/index'
 
 
 
-export default async function () {
+export default async function (createAppFn) {
   // create store and router instances
   
   const store = typeof createStore === 'function'
-    ? await createStore({Vue})
+    ? await createStore({})
     : createStore
   
   const router = typeof createRouter === 'function'
-    ? await createRouter({Vue, store})
+    ? await createRouter({store})
     : createRouter
   
   // make router instance available in store
   store.$router = router
   
 
-  // Create the app instantiation Object.
-  // Here we inject the router, store to all child components,
-  // making them available everywhere as `this.$router` and `this.$store`.
-  const app = {
-    router,
-    store,
-    render: h => h(App)
-  }
-
+  // Create the app instance.
+  // Here we inject into it the Quasar UI, the router & possibly the store.
+  const app = createAppFn(RootComponent)
 
   
-  app.el = '#q-app'
+  app.config.devtools = true
   
 
-  // expose the app, the router and the store.
-  // note we are not mounting the app here, since bootstrapping will be
+  app.use(Quasar, quasarUserOptions)
+
+  
+
+  // Expose the app, the router and the store.
+  // Note that we are not mounting the app here, since bootstrapping will be
   // different depending on whether we are in a browser or on the server.
   return {
     app,
