@@ -5,17 +5,20 @@
 
 // Configuration for your app
 // https://quasar.dev/quasar-cli/quasar-conf-js
+
 /* eslint-env node */
 /* eslint-disable @typescript-eslint/no-var-requires */
-/* eslint-disable @typescript-eslint/camelcase */
 const { configure } = require('quasar/wrappers');
 
-module.exports = configure(function(ctx) {
+module.exports = configure(function (ctx) {
   return {
     // https://quasar.dev/quasar-cli/supporting-ts
     supportTS: {
       tsCheckerConfig: {
-        eslint: true
+        eslint: {
+          enabled: true,
+          files: './src/**/*.{ts,tsx,js,jsx,vue}',
+        },
       }
     },
 
@@ -25,10 +28,14 @@ module.exports = configure(function(ctx) {
     // app boot file (/src/boot)
     // --> boot files are part of "main.js"
     // https://quasar.dev/quasar-cli/boot-files
-    boot: ['notify-defaults', 'guard', 'apollo'],
+    boot: [
+      'composition-api', 'notify-defaults'
+    ],
 
     // https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-css
-    css: ['app.scss'],
+    css: [
+      'app.scss', 'mainLayout.scss'
+    ],
 
     // https://github.com/quasarframework/quasar/tree/dev/extras
     extras: [
@@ -41,12 +48,13 @@ module.exports = configure(function(ctx) {
       // 'roboto-font-latin-ext', // this or either 'roboto-font', NEVER both!
 
       'roboto-font', // optional, you are not bound to it
-      'material-icons' // optional, you are not bound to it
+      'material-icons', // optional, you are not bound to it
     ],
 
     // Full list of options: https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-build
     build: {
       vueRouterMode: 'hash', // available values: 'hash', 'history'
+
       // transpile: false,
 
       // Add dependencies for transpiling with Babel (Array of string/regex)
@@ -64,9 +72,9 @@ module.exports = configure(function(ctx) {
       // extractCSS: false,
 
       // https://quasar.dev/quasar-cli/handling-webpack
+      // "chain" is a webpack-chain object https://github.com/neutrinojs/webpack-chain
       extendWebpack(cfg) {
         // linting is slow in TS projects, we execute it only for production builds
-        // cfg.externals = { 'fs-extra': 'commonjs2 fs-extra' };
         if (ctx.prod) {
           if (process.env.NODE_ENV === 'production') {
             cfg.module.rules.push({
@@ -79,7 +87,7 @@ module.exports = configure(function(ctx) {
         }
       },
 
-      chainWebpack(chain, { isServer, isClient }) {
+      chainWebpack(chain) {
         chain.module
           .rule('vue')
           .use('vue-loader')
@@ -92,7 +100,8 @@ module.exports = configure(function(ctx) {
             };
             return options;
           });
-      }
+      },
+
     },
 
     // Full list of options: https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-devServer
@@ -104,14 +113,19 @@ module.exports = configure(function(ctx) {
 
     // https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-framework
     framework: {
-      iconSet: 'material-icons', // Quasar icon set
-      lang: 'en-us', // Quasar language pack
-      config: {},
+      config: {
+        notify: { /* look at QuasarConfOptions from the API card */ }
+      },
 
-      // Possible values for "importStrategy":
-      // * 'auto' - (DEFAULT) Auto-import needed Quasar components & directives
-      // * 'all'  - Manually specify what to import
-      importStrategy: 'auto',
+      // iconSet: 'material-icons', // Quasar icon set
+      lang: 'en-US', // Quasar language pack
+
+      // For special cases outside of where the auto-import stategy can have an impact
+      // (like functional components as one of the examples),
+      // you can manually specify Quasar components/directives to be available everywhere:
+      //
+      // components: [],
+      // directives: [],
 
       // Quasar plugins
       plugins: ['Notify']
@@ -132,11 +146,11 @@ module.exports = configure(function(ctx) {
       workboxOptions: {}, // only for GenerateSW
       manifest: {
         name: 'Quasar App',
-        short_name: 'QuasarApp',
+        short_name: 'Quasar App',
         description: 'A Quasar Framework app',
         display: 'standalone',
         orientation: 'portrait',
-        backgroundColor: '#ffffff',
+        background_color: '#ffffff',
         theme_color: '#027be3',
         icons: [
           {
@@ -169,6 +183,14 @@ module.exports = configure(function(ctx) {
     },
 
     // Full list of options: https://quasar.dev/quasar-cli/developing-cordova-apps/configuring-cordova
+    cordova: {
+      // noIosLegacyBuildFlag: true, // uncomment only if you know what you are doing
+    },
+
+    // Full list of options: https://quasar.dev/quasar-cli/developing-capacitor-apps/configuring-capacitor
+    capacitor: {
+      hideSplashscreen: true
+    },
 
     sourceFiles: {
       electronMainDev: 'src-electron/main-process/electron-main.dev.ts',
@@ -181,11 +203,13 @@ module.exports = configure(function(ctx) {
 
       packager: {
         // https://github.com/electron-userland/electron-packager/blob/master/docs/api.md#options
+
         // OS X / Mac App Store
         // appBundleId: '',
         // appCategoryType: '',
         // osxSign: '',
         // protocol: 'myapp://path',
+
         // Windows only
         // win32metadata: { ... }
       },
@@ -194,7 +218,7 @@ module.exports = configure(function(ctx) {
         appId: 'two_manager',
         // https://www.electron.build/configuration/configuration
         compression: 'maximum',
-        appId: 'update_version',
+        appId: 'twom',
         directories: {
           output: 'dist',
           buildResources: 'build'
@@ -252,5 +276,5 @@ module.exports = configure(function(ctx) {
         });
       }
     }
-  };
+  }
 });
