@@ -1,77 +1,32 @@
 <template>
-  <q-card class="  columns" style="padding: 15px">
-    <q-card-section class="col-1 text-center">
-      <p class="text-bold">Ajouter un Client</p>
-    </q-card-section>
-
-    <q-card-section class=" q-gutter-lg col-11">
-      <q-input
-        dense
-        rounded
-        outlined
-        v-model="clientNameState"
-        label="Nom complet *"
-        :rules="[
-          val => (val && val.length > 0) || 'Ce champ ne doit pas être vide.'
-        ]"
-      />
-
-      <q-input
-        dense
-        rounded
-        outlined
-        v-model="domicileState"
-        label="Domicile"
-      />
-
-      <q-input
-        dense
-        rounded
-        outlined
-        v-model="contactState"
-        mask="### ## ### ##"
-        fill-mask="*"
-        label="Contact"
-      />
-      <q-btn
-        rounded
-        label="Valider"
-        type="submit"
-        color="amber"
-        text-color="black"
-      />
-      <q-btn
-        rounded
-        label="Reset"
-        color="orange"
-        text-color="black"
-        @click="
-          clientNameState = '';
-          domicileState = '';
-          contactState = '';
-        "
-      />
-    </q-card-section>
-  </q-card>
+  <div style="position: relative; padding-left: 10px">
+    <ClientAddForm v-if="clientFormBtn === 'add'" />
+    <ClientUpdateForm v-if="clientFormBtn === 'update'" />
+  </div>
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-import { mapFields } from 'vuex-map-fields';
+import { defineComponent, ref, Ref } from 'vue';
+import { useAddClientByUser } from '../../services/clients/useAddClientByUser';
 
-@Component({
-  name: 'ClientForm',
-  computed: {
-    ...mapFields({
-      clientNameState: 'clientsModule.form.clientName',
-      domicileState: 'clientsModule.form.domicile',
-      contactState: 'clientsModule.form.contact'
-    })
-  }
-})
-export default class ClientForm extends Vue {
-  private clientNameState: string;
-  private domicileState: string;
-  private contactState: string;
-}
+type ClientFormBtn = null | 'add' | 'update';
+
+export const clientFormBtn: Ref<ClientFormBtn> = ref();
+
+export default defineComponent({
+  name: 'clientForm',
+  components: {
+    ClientAddForm: require('./ClientForm.add.vue').default,
+    ClientUpdateForm: require('./ClientForm.update.vue').default,
+  },
+  setup: () => {
+    const [state, submitAddClient] = useAddClientByUser();
+
+    return {
+      state,
+      submitAddClient,
+      clientFormBtn,
+    };
+  },
+});
 </script>
