@@ -7,18 +7,18 @@
       @click="slide()"
       :width="width"
     >
-      <div class=" row justify-center items-center full-height full-width">
-        <q-img src="~assets/home4.jpg" style="width: 100%; height: 100%;">
+      <div class="row justify-center items-center full-height full-width">
+        <q-img src="~assets/home4.jpg" style="width: 100%; height: 100%">
           <div
-            class=" full-width full-height text-center"
-            style="background: #c3870075; width: 100%; height: 100%;"
+            class="full-width full-height text-center"
+            style="background: #c3870075; width: 100%; height: 100%"
           ></div>
         </q-img>
         <h4
           :style="
             'position: absolute; font-weight: bold; color: white; font-size: ' +
-              size +
-              '%'
+            size +
+            '%'
           "
         >
           S'enregistrer
@@ -34,12 +34,12 @@
       :width="width"
     >
       <div
-        class=" row justify-center items-center full-height full-width boredr"
+        class="row justify-center items-center full-height full-width boredr"
       >
         <q-img dark src="~assets/home1.jpg" style="width: 100%; height: 100%">
           <div
-            class=" full-width full-height text-center"
-            style="background: #ff980054; width: 100%; height: 100%;"
+            class="full-width full-height text-center"
+            style="background: #ff980054; width: 100%; height: 100%"
           ></div>
         </q-img>
 
@@ -58,17 +58,16 @@
         :myclass="{
           form: 'form-black',
           h1: 'h1-black',
-          dark: true
+          dark: true,
         }"
-        style="width: 500px;
-      height: 600px"
+        style="width: 500px; height: 600px"
       />
       <Register
         v-else
         :myclass="{
           form: 'form-black',
           h1: 'h1-black',
-          dark: true
+          dark: true,
         }"
         style="width: 500px; height: 600px"
       />
@@ -77,34 +76,33 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
+import { defineComponent, ref } from 'vue';
 
-@Component({
-  name: 'Authentification',
+export default defineComponent({
   components: {
     Login: require('src/components/authentification/Login.vue').default,
-    Register: require('src/components/authentification/Register.vue').default
-  }
-})
-export default class Index extends Vue {
-  private left = true;
-  private right = false;
-  private width = 0;
-  private size = 0;
+    Register: require('src/components/authentification/Register.vue').default,
+  },
+  setup: () => {
+    const left = ref(true);
+    const right = ref(false);
+    const width = ref(0);
+    const size = ref(0);
 
-  created() {
-    window.addEventListener('resize', () => this.handleResize());
-    this.handleResize();
-  }
+    function handleResize() {
+      width.value = (35 * window.innerWidth) / 100;
+      size.value = (25 * window.innerWidth) / 100;
+    }
 
-  handleResize() {
-    this.width = (35 * window.innerWidth) / 100;
-    this.size = (25 * window.innerWidth) / 100;
-  }
+    function slide() {
+      left.value = !left.value;
+      right.value = !right.value;
+    }
 
-  slide() {
-    this.left = !this.left;
-    this.right = !this.right;
-  }
-}
+    window.addEventListener('resize', () => handleResize());
+    handleResize();
+
+    return { left, right, width, size, slide };
+  },
+});
 </script>
