@@ -1,40 +1,21 @@
 <template>
-  <q-page class="row items-center justify-center">
-    <div class=" " style="width: 300px; height: 500px">{{}}</div>
+  <q-page class="">
+    <div class="q-pa-md" style="min-height: 720px; max-height: 900px">
+      <ClientTable />
+    </div>
   </q-page>
 </template>
 
 <script lang="ts">
-import {
-  ref,
-  computed,
-  watch,
-  toRef,
-  reactive,
-  defineComponent
-} from '@vue/composition-api';
-
-import { mapFields } from 'vuex-map-fields';
-import { Accessors } from 'types';
-
-import { LoginForm } from 'src/store/users/login/login.types';
+import { defineComponent } from 'vue';
 
 export default defineComponent({
-  computed: {
-    ...(mapFields({
-      login: 'usersModule.loginState.form.login',
-      password: 'usersModule.loginState.form.password'
-    }) as Accessors<LoginForm>)
+  name: 'client',
+  components: {
+    ClientTable: require('src/components/client/ClientTable.vue').default,
   },
   setup: () => {
-    const blabla = 'string';
-
-    const g = computed(() => ({}));
-
-    return { blabla };
+    return {};
   },
-  methods: {
-    test12: () => this.blabla
-  }
 });
 </script>
