@@ -1,5 +1,6 @@
-import { Notify } from 'quasar';
+import { useQuasar, Notify, QVueGlobals } from 'quasar';
 import { GraphQLError } from 'graphql/error/GraphQLError';
+
 
 export const context = () => ({
   headers: {
@@ -7,14 +8,25 @@ export const context = () => ({
   }
 });
 
-export const notifyThis = (message: string, type = 'warning') => {
+type MessageType = 'positive' | 'negative' | 'warning' | 'info';
+type Position = 'bottom-right' | 'top-left' | 'top-right' | 'bottom-left' | 'top' | 'bottom' | 'left' | 'right' | 'center'
+interface ThisNotify {
+  message: string,
+  type: MessageType,
+  position?: Position,
+  timeout?: number,
+  progress?: boolean
+}
+
+export const notifyThis = ({message, type, position = 'bottom-right', timeout = 6000, progress = true }: ThisNotify) => {
+
   Notify.create({
-    type: type,
+    type,
     message,
-    position: 'bottom-right',
-    timeout: 6000,
+    position,
+    timeout,
     multiLine: true,
-    progress: true,
+    progress,
     actions: [
       {
         label: 'Compris',
@@ -27,15 +39,17 @@ export const notifyThis = (message: string, type = 'warning') => {
   });
 };
 
+
 export const notifyThere = (
   errors: readonly GraphQLError[],
-  type = 'warning'
+  type: MessageType = 'warning'
 ) => {
   let i = 0;
   errors.map(err => {
+    // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
     const time = parseInt(i++ + '000', 10);
     setTimeout(() => {
-      notifyThis(err.message, type);
+      notifyThis({message: err.message, type});
     }, time);
   });
 };
