@@ -1,7 +1,6 @@
 // Mocks all files ending in `.vue` showing them as plain Vue instances
 declare module '*.vue' {
-  import Vue from 'vue';
-  export default Vue;
+  import { ComponentOptions } from 'vue';
+  const component: ComponentOptions;
+  export default component;
 }
-
-declare module 'http' {}
